+++ frontend/src/components/screens/departments/hhp/technicians/page.tsx
@@ -103,6 +103,7 @@ import repairshopr_statuses_techs from '@/lib/tech_rs_statuses'
 import moment from 'moment'
 import openEngineerBinsTab from "@/lib/openEngineerBinsTab"
 import useAddCommentsLocally from "@/hooks/useCommentsLocally"
+import BinsScreen from "../bins/page"
 const DateCalculationsScreen = dynamic(() =>
     import('./date_calculations/page'), { ssr: false }
 )
@@ -139,7 +140,7 @@ const TechniciansScreen = () => {
     const { updateAssessmentDate } = useUpdateAssessmentDate()
     const [dateFrom, setDateFrom] = useState("")
     const [dateTo, setDateTo] = useState("")
- 
+
 
 
     // engineer filters
@@ -180,7 +181,7 @@ const TechniciansScreen = () => {
     const [search_part, setSearchPart] = useState("")
 
     const { engineersList } = useFetchEngineer();
-    const hhpTechs = engineersList?.filter((x)=> x.department === "HHP")
+    const hhpTechs = engineersList?.filter((x) => x.department === "HHP")
     // for filtering by engineer
     const engineerListFomatted = hhpTechs?.map((user) => ({
         id: user?.id,
@@ -347,192 +348,30 @@ const TechniciansScreen = () => {
 
                             <PageTitle title="Management" hasSpan={true} spanText={"HHP"} />
 
-                            {/* modal for sorting table columns */}
-                            {
-                                openSortTableColumnsModal &&
-                                <Dialog open={openSortTableColumnsModal} onOpenChange={() => setSortTableColumns(false)} >
-                                    {/* <DialogTrigger>Open</DialogTrigger> */}
-                                    <DialogContent>
-                                        <DialogHeader>
-                                            <DialogTitle>Sort columns</DialogTitle>
-                                            <DialogDescription>
-                                                Toggle columns you want/do not want
-                                            </DialogDescription>
-                                        </DialogHeader>
-
-                                        <div className="inline-block border border-black shadow rounded">
-                                            <div className="px-1 border-b border-black">
-                                                <label>
-                                                    <input
-                                                        className='bg-sky-600 cursor-pointer checked:bg-slate-950 focus:bg-slate-950'
-                                                        {...{
-                                                            type: 'checkbox',
-                                                            checked: table.getIsAllColumnsVisible(),
-                                                            onChange: table.getToggleAllColumnsVisibilityHandler(),
-                                                        }}
-                                                    />{' '}
-                                                    Show them all
-                                                </label>
-                                            </div>
-                                            {table.getAllLeafColumns().map(column => {
-                                                return (
-                                                    <div key={column.id} className="px-1">
-                                                        <label className='text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-                                                            <input
-                                                                className='bg-sky-600 cursor-pointer checked:bg-slate-950 focus:bg-slate-950'
-                                                                {...{
-                                                                    type: 'checkbox',
-                                                                    checked: column.getIsVisible(),
-                                                                    onChange: column.getToggleVisibilityHandler(),
-                                                                }}
-                                                            />{' '}
-                                                            {column?.columnDef?.header as any}
-
-                                                        </label>
-                                                    </div>
-                                                )
-                                            })}
-                                        </div>
-                                    </DialogContent>
-                                </Dialog>
-                            }
-                            <section className="flex flex-wrap justify-between items-center py-5 gap-2 w-full overflow-auto">
-                                {/* Search Field - Allow it to grow */}
-                                <div className="flex-grow min-w-[200px]">
-                                    <ManagementSearchForm
-                                        filtering={filtering}
-                                        setFiltering={(e) => setFiltering(e.target.value)}
-                                    />
+                            <div className="grid lg:grid-cols-2 grid-cols-1 items-center gap-3">
+                                <div className="h-full border">
+                                    <table>
+                                        <tr>
+                                            <th>Company</th>
+                                            <th>Contact</th>
+                                            <th>Country</th>
+                                        </tr>
+                                        <tr>
+                                            <td>Alfreds Futterkiste</td>
+                                            <td>Maria Anders</td>
+                                            <td>Germany</td>
+                                        </tr>
+                                        <tr>
+                                            <td>Centro comercial Moctezuma</td>
+                                            <td>Francisco Chang</td>
+                                            <td>Mexico</td>
+                                        </tr>
+                                    </table>
                                 </div>
-                                <div className="flex justify-between items-center gap-3">
-
-                                    {
-                                        user?.user_role === "admin" ?
-                                            <>
-
-                                                <Button type="button" onClick={() => openEngineerBinsTab('/departments/hhp/bins')}>Bin stats</Button>
-                                                <Button type="button" onClick={downloadReport} disabled={reportsLoading}>{reportsLoading ? 'Downloading...' : 'Get report'}</Button>
-                                            </>
-                                            : null
-                                    }
-                                    {
-                                        user?.user_role === "admin" ?
-                                            <div className="flex gap-3 items-center">
-                                                <span>
-                                                    <Label
-                                                        htmlFor="dateFrom"
-                                                        className="sr-only dark:text-[#eee]"
-                                                    >
-                                                        Date from
-                                                    </Label>
-                                                    <Input
-                                                        type="date"
-                                                        name="dateFrom"
-                                                        value={dateFrom}
-                                                        onChange={handleDateFromFilter}
-                                                        className="cursor-pointer"
-                                                        id="dateFrom"
-                                                    />
-                                                </span>
-                                                <span>-</span>
-                                                <span>
-                                                    <Label
-                                                        htmlFor="dateTo"
-                                                        className="sr-only dark:text-[#eee]"
-                                                    >
-                                                        Date to
-                                                    </Label>
-                                                    <Input
-                                                        type="date"
-                                                        name="dateTo"
-                                                        value={dateTo}
-                                                        onChange={handleDateToFilter}
-                                                        className="cursor-pointer"
-                                                        id="dateTo"
-                                                    />
-                                                </span>
-                                            </div> : null
-                                    }
-                                    <Select name="statusFilter" value={statusFilter} onValueChange={handleStatusFilter}>
-                                        <SelectTrigger className="w-full hidden md:flex">
-                                            <SelectValue placeholder="Status" />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {isLoggedIn && user?.user_role === "admin" ?
-                                                <SelectGroup>
-                                                    <SelectLabel>Status</SelectLabel>
-                                                    {repairshopr_statuses.map((dep) => (
-                                                        <SelectItem key={dep.id} value={`${dep._status}`}>{`${dep._status}`}</SelectItem>
-                                                    ))}
-                                                </SelectGroup> :
-                                                <SelectGroup>
-                                                    <SelectLabel>Status</SelectLabel>
-                                                    {repairshopr_statuses_techs.map((dep) => (
-                                                        <SelectItem key={dep.id} value={`${dep._status}`}>{`${dep._status}`}</SelectItem>
-                                                    ))}
-                                                </SelectGroup>}
-                                        </SelectContent>
-                                    </Select>
-
-                                    <Select name="engineerFilter" value={unassisgnedFilter} onValueChange={handleUnassignedFilter}>
-                                        <SelectTrigger className="w-full hidden md:flex">
-                                            <SelectValue placeholder="Unassigned" />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            <SelectItem value={"unassigned"}>Unassigned</SelectItem>
-                                        </SelectContent>
-                                    </Select>
-                                    <Select name="engineerFilter" value={engineerFilter} onValueChange={handleEngineerFilter}>
-                                        <SelectTrigger className="w-full hidden md:flex">
-                                            <SelectValue placeholder="Technician" />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            <SelectGroup>
-                                                <SelectLabel>Engineer</SelectLabel>
-                                                {engineerListFomatted.map((dep) => (
-                                                    <SelectItem key={dep.id} value={`${dep.value}`}>{`${dep.label}`}</SelectItem>
-                                                ))}
-                                            </SelectGroup>
-                                        </SelectContent>
-                                    </Select>
-                                    <Button type="button" onClick={resetFilters}> Reset filters</Button>
-                                    <Button type="button" onClick={() => setSortTableColumns(true)} className="hidden md:block">Sort columns</Button>
-                                    <Button type="button" onClick={() => setOpenAddTaskModal(true)}> Add task</Button>
+                                <div className="h-full border">
+                                    <BinsScreen />
                                 </div>
-
-                            </section>
-                            <div className="overflow-y-auto max-h-[540px] rounded-lg shadow-lg">
-                                <table className="w-full whitespace-nowrap text-sm text-left text-gray-500 table-auto">
-                                    <TableHead table={table} />
-                                    <TableBody table={table} handleRowClick={handleRowClick} deleteRow={handleDeleteRow} handleOpenSinglePage={handleOpenSinglePage} />
-                                </table>
                             </div>
-                            <div className="h-2" />
-                            <Pagination table={table} />
-
-
-                            {/* modal for adding task */}
-                            {
-                                openAddTaskModal &&
-                                <Modal
-                                    isVisible={openAddTaskModal}
-                                    onClose={() => setOpenAddTaskModal(false)}
-                                    title={"Add task"}
-                                    content={
-                                        <Tabs defaultValue="repairshopr">
-                                            <TabsList>
-                                                <TabsTrigger value="gspn">GSPN</TabsTrigger>
-                                                <TabsTrigger value="repairshopr">Repairshopr</TabsTrigger>
-                                            </TabsList>
-                                            <TabsContent value="gspn"><AddgspnHHPTask onChange={() => setOpenAddTaskModal(false)} /></TabsContent>
-                                            <TabsContent value="repairshopr"><AddRepairshoprHHPTask onChange={() => setOpenAddTaskModal(false)} /></TabsContent>
-                                        </Tabs>
-
-                                    }
-                                />
-                            }
-
-
                         </main>
                     </>
                 ) : (