+++ server/controllers/department/claims/create_claim.js
@@ -52,6 +52,7 @@ const CreateClaim = async (req, res) => {
             });
         }
     } catch (error) {
+        console.log(error);
         // Handle validation or other errors
         const errors = {};
         if (error.inner) {