+++ server/routes/department/hhp/hhp_jobs_route.js
@@ -13,11 +13,12 @@ import { UpdateTask } from "../../../controllers/department/hhp/technicians/upda
 import { limiter } from "../../../middleware/rateLimiter.js";
 import { authenticateToken } from "../../../middleware/verify.js";
 import { authenticateAdmin } from "../../../middleware/verify_admin.js";
+import AddToQCTable from "../../../controllers/department/hhp/technicians/add_to_qc_table.js";
 const router = express.Router();
 
 router.post("/", limiter, authenticateToken, AddHHPTask);
 router.patch("/assess/:id", limiter, authenticateToken, UpdateAssessmentDate);
-
+router.post("/qc_table", AddToQCTable);
 router.get("/", authenticateToken, GetAllTasks);
 router.get("/:id", authenticateToken, GetTaskById);
 router.get("/engineer/bin", authenticateAdmin, GetBinStats);
@@ -27,4 +28,3 @@ router.patch("/so/:id", authenticateToken, UpdateSOAfterBooking);
 router.delete("/:id", authenticateAdmin, deleteHHPTask);
 
 export { router };
-